@@ -1,14 +1,35 @@
 {
-  "name": "my-ppp",
+  "name": "nos_ik_proj",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
+    "@capacitor/device": "^4.0.1",
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/lab": "^5.0.0-alpha.94",
+    "@mui/material": "^5.10.0",
+    "@mui/styled-engine-sc": "^5.10.0",
+    "@reduxjs/toolkit": "^1.8.3",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
+    "apexcharts": "^3.35.4",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.2.0",
+    "moment": "^2.29.4",
+    "react": "^18.0.0",
+    "react-bootstrap": "^2.4.0",
     "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "react-file-base64": "^1.0.3",
+    "react-helmet": "^6.1.0",
+    "react-icons": "^4.4.0",
+    "react-redux": "^8.0.2",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^2.1.3",
+    "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
