@@ -1,15 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
- 
+import React from "react";
+import { BrowserRouter as Router } from "react-router-dom";
 
- 
-ReactDOM.render(< App /> , document.getElementById('root'))
+import App from "./App";
+import { store } from "./features/store";
+import { Provider } from "react-redux";
+import { createRoot } from "react-dom/client";
 
-
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
- 
+const root = createRoot(document.getElementById("root"));
+root.render(
+  <Router>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </Router>,
+  root
+);
